+++ e2e/tests/s3/mod.rs
@@ -8,7 +8,11 @@ use baza_api_s3 as s3;
 use cucumber::{gherkin::Step, given, then, when};
 use rusoto_core::{region::Region, HttpClient, RusotoError};
 use rusoto_credential::StaticProvider;
-use rusoto_s3::{PutObjectError, PutObjectRequest, S3Client, S3 as _};
+use rusoto_s3::{
+    GetObjectError, GetObjectRequest, PutObjectError, PutObjectRequest,
+    S3Client, S3 as _,
+};
+use tokio::io::AsyncReadExt as _;
 
 use super::{sample_file, World, DATA_DIR};
 
@@ -30,6 +34,11 @@ async fn file_uploaded(
         None::<String>,
     )
     .await
+}
+
+#[given(regex = r"^there was nothing uploaded to `(\S+)` bucket as `(\S+)`$")]
+async fn nothing_uploaded(_: &mut World) {
+    // nothing is uploaded by default
 }
 
 #[given(regex = "^`(\\S+)` symlink was created on `(\\S+)` bucket \
@@ -100,13 +109,38 @@ async fn invalid_argument_error(w: &mut World) {
         .await;
 }
 
+#[when(regex = r"^trying to load `(\S+)` from `(\S+)` bucket$")]
+async fn trying_to_load_file(w: &mut World, key: String, bucket: String) {
+    w.get_object_response =
+        Some(try_get_object(bucket, w.unique.filename(key)).await);
+}
+
+#[then(regex = r"^`(\S+)` file is returned$")]
+async fn file_is_returned(w: &mut World, name: String) {
+    let sample = sample_file(name);
+
+    let file = w
+        .last_get_object_response()
+        .unwrap_or_else(|e| panic!("`GetObjectRequest` failed: {e}"));
+
+    assert_eq!(sample.len(), file.len());
+    assert!(sample == file, "Bytes don't match");
+}
+
+#[then(regex = r"^`NoSuchKey` error is returned$")]
+async fn error_is_returned(w: &mut World) {
+    let res = w.last_get_object_response();
+    match res {
+        Err(RusotoError::Service(GetObjectError::NoSuchKey(_))) => {}
+        _ => panic!("Expected `NoSuchKey` error, got: {res:#?}"),
+    }
+}
+
 fn assert_invalid_argument(res: Result<(), RusotoError<PutObjectError>>) {
     match &res {
         Err(RusotoError::Unknown(resp))
             if resp.body_as_str().contains("InvalidArgument") => {}
-        _ => {
-            panic!("Expected InvalidArgument error, got: {res:#?}");
-        }
+        _ => panic!("Expected `InvalidArgument` error, got: {res:#?}"),
     }
 }
 
@@ -118,8 +152,12 @@ async fn put_object(
 ) {
     try_put_object(bucket, key, body, symlink_to)
         .await
-        .unwrap_or_else(|e| panic!("PutObjectRequest failed: {}", e));
+        .unwrap_or_else(|e| panic!("`PutObjectRequest` failed: {e}"));
 }
+
+/// Response to a [`GetObjectRequest`].
+pub(super) type GetObjectResponse =
+    Result<Vec<u8>, RusotoError<GetObjectError>>;
 
 async fn try_put_object(
     bucket: impl ToString,
@@ -142,15 +180,50 @@ async fn try_put_object(
     s3_client().put_object(req).await.map(drop)
 }
 
+async fn try_get_object(
+    bucket: impl ToString,
+    key: impl ToString,
+) -> GetObjectResponse {
+    let req = GetObjectRequest {
+        bucket: bucket.to_string(),
+        key: key.to_string(),
+        ..GetObjectRequest::default()
+    };
+
+    let resp = s3_client().get_object(req).await?;
+    let mut buf = Vec::new();
+    resp.body
+        .unwrap()
+        .into_async_read()
+        .read_to_end(&mut buf)
+        .await
+        .unwrap();
+
+    Ok(buf)
+}
+
 /// Creates a new [`S3Client`] for performing requests to the S3 HTTP API being
 /// tested.
 fn s3_client() -> S3Client {
     S3Client::new_with(
         HttpClient::new().expect("Failed to initialize Rusoto HTTP client"),
-        StaticProvider::new_minimal("test".into(), "test".into()),
+        StaticProvider::new_minimal("baza".into(), "baza".into()),
         Region::Custom {
             name: "test".into(),
             endpoint: API_URL.into(),
         },
     )
+}
+
+impl World {
+    /// Takes the last [`GetObjectResponse`], stored in this [`World`].
+    ///
+    /// # Panics
+    ///
+    /// If there is no [`GetObjectResponse`] in this [`World`].
+    fn last_get_object_response(&mut self) -> GetObjectResponse {
+        self.get_object_response
+            .take()
+            .expect("No `GetObjectResponse`")
+    }
 }